Create Custom Cipher.py
This builds on the simplified version of GGHYK-M and adds a scrambling algorithm, which requires almost the exact matrix values for encryption.  From the tests run, if all of the values are the same but one value within 5+/- in a 2x2 matrix, it will result in the correct output, but 6+/- results in an unreadable scrambled text.  

This version added a scrambling algorithm to closely require the exact matrix used for encryption to be used for decryption.